+++ src/App.vue
@@ -57,6 +57,16 @@ export default {
                 initializeLayers(this.visibleLayerConfigs);
                 startProcessUrlParams();
                 this.initializeOther();
+
+                // Wait until next tick to ensure that the menu components is rendered
+                await this.$nextTick();
+
+                // Check if login module is available after configs are loaded
+                if (this.$store?.getters?.isModuleAvailable?.("login")) {
+                    // Start periodic check to verify if user token is still valid
+                    await this.setUpTokenRefreshInterval();
+                }
+
             }
         }
     },
@@ -86,6 +96,7 @@ export default {
             "loadRestServicesJson",
             "loadServicesJson"
         ]),
+        ...mapActions("Modules/Login", ["checkLoggedIn", "setUpTokenRefreshInterval"]),
 
         /**
          * Sets global variables.