+++ src/modules/login/store/actionsLogin.js
@@ -3,6 +3,20 @@ import OIDC from "../js/utilsOIDC";
 
 export default {
     /**
+     * Starts a timer to periodically check if the user is still logged in
+     * @param {Object} context the context Vue instance
+     * @return {void}
+     */
+    async setUpTokenRefreshInterval ({dispatch}) {
+        // Initial check
+        await dispatch("checkLoggedIn");
+        // Set up interval periodically check
+        setInterval(() => {
+            dispatch("checkLoggedIn");
+        }, 10_000);
+    },
+
+    /**
      * Returns authentication URL
      *
      * @param {Object} context the context Vue instance